Simplified login and signup codes, and varaible names, and edited Nav Bar so that it is more pleasing
All logics are good. One thing to remember is that we don't have to check the signin email format, because if it is not correct, they can't sign in. In addition, currently, the login error message is that if you don't enter password, it reads the password is not matching, but this is just to check if the error message works. What really checks that is after backend data can save the password. Finally, I also simplified certain varaible names so that it would be more readable. I have not been able to connect this with the API json file, since I can't find it in the directory. However, it should be easy for backend engineers to read this file, as it is very clear now. I can also do a quick walkthrough, if anyone has questions about this frontend codes.