Signup login homepage week8 (#120)
* Basic signup container and fields working

* Added single sign on with Google

* Added all login and signup layout code

Most of the layout is done. The functionality is almost done, a few bugs exist in alerts and setting password and email rules. These logics will be done tomorrow.

* Layout and basic function working

Some fine tuning could include adding Userfront API to implement the login and signup function. Other things to consider include nav bar.

* Added Nav bar

* Added Nav bar. All layout complete

* Simplified login and signup codes, and varaible names, and edited Nav Bar so that it is more pleasing

All logics are good. One thing to remember is that we don't have to check the signin email format, because if it is not correct, they can't sign in. In addition, currently, the login error message is that if you don't enter password, it reads the password is not matching, but this is just to check if the error message works. What really checks that is after backend data can save the password. Finally, I also simplified certain varaible names so that it would be more readable. I have not been able to connect this with the API json file, since I can't find it in the directory. However, it should be easy for backend engineers to read this file, as it is very clear now. I can also do a quick walkthrough, if anyone has questions about this frontend codes.

* Added eslint and see if it will pass travis

* Seperated all components in homepage

Next step is to add switch function so that the users can choose if they want to sign up or log in

* Deleted some redundant css

* Added Sign up and Log in Siwtch function so that we have seperate views

Note on how to approach the files: All most updated working files are in switchViews directory, and they are linked properly to things in DemoView, LoginView, and SignupView in views folder.

* Changed Navbar fonts

* checking linter

* Solved a bunch of conflicts and linter problems

* Linter solved

* Linter works locally

* Delete all userfront and retried linter

* Linter solved and deleted redundant files

* Changed the file names in signup_login directory to reduce confusion

* Used the old cjs and kept the new js for eslint.

* Deleted eslintrc.cjs

* Deleted cjs file and changed the json files

* Changed json file and eslint file accordingly